requirements.txt
discord.py==1.5.1
Pillow==8.0.1
pymongo[srv]==3.11.2